+++ src/mining/miner.h
@@ -207,7 +207,7 @@ private:
     /** Add a tx to the block */
     void AddToBlock(MCTxMemPool::txiter iter, MakeBranchTxUTXO& utxoMaker);
 
-    void GroupingTransaction(int offset, std::vector<const MCTxMemPoolEntry*>& blockTxEntries);
+    void GroupingTransaction(size_t offset, std::vector<const MCTxMemPoolEntry*>& blockTxEntries);
 
     // Methods for how to add transactions to a block.
     /** Add transactions based on feerate including unconfirmed ancestors
